backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.4.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
alembic>=1.12.1
httpx>=0.25.1
sentence-transformers>=2.2.2
chromadb>=0.4.18
aiosqlite>=0.19.0
python-dotenv>=1.0.0
